[2607-DEV-482] Collapse TripCard and SocialsTile dual-layout files into single responsive components
Neither TripCardDesktop/Mobile nor SocialsTileDesktop/Mobile hit any of the
four documented triggers for a dual-layout exception (.cursor/rules/frontend.mdc:19-25):
no 5+ column table, no drag-and-drop, no persistent sidebar, no canvas/map/editor.
Merges each pair into a single file using Tailwind md: modifiers per the rule's
own prescribed default, and updates the two call sites.

- app/(dashboard)/trips/components/TripCard.tsx (new): merges TripCardDesktop.tsx
  + TripCardMobile.tsx; TripImage's height prop generalized to a responsive
  heightClassName since desktop/mobile used different fixed heights.
- app/(dashboard)/components/tiles/SocialsTile.tsx: replaces a stale, unreferenced
  pre-split version with the real merge of SocialsTileDesktop.tsx + SocialsTileMobile.tsx.
- app/(dashboard)/trips/TripsClient.tsx, app/(dashboard)/page.tsx: call sites updated.
- Deletes the four now-redundant Desktop/Mobile files.

Closes #482
